update specs in Godeps/
