--- Godeps/_workspace/src/github.com/opencontainers/specs/MAINTAINERS
@@ -1,8 +1,0 @@
-Michael Crosby <michael@docker.com> (@crosbymichael)
-Alexander Morozov <lk4d4@docker.com> (@LK4D4)
-Vishnu Kannan <vishnuk@google.com> (@vishnuk)
-Mrunal Patel <mpatel@redhat.com> (@mrunalp)
-Vincent Batts <vbatts@redhat.com> (@vbatts)
-Daniel, Dao Quang Minh <dqminh89@gmail.com> (@dqminh)
-Brandon Philips <brandon.philips@coreos.com> (@philips)
-Tianon Gravi <admwiggin@gmail.com> (@tianon)
