--- Godeps/_workspace/src/github.com/opencontainers/specs/implementations.md
@@ -1,21 +1,0 @@
-# Implementations
-
-The following sections link to associated projects, some of which are maintained by the OCI and some of which are maintained by external organizations.
-If you know of any associated projects that are not listed here, please file a pull request adding a link to that project.
-
-## Runtime (Container)
-
-* [opencontainers/runc](https://github.com/opencontainers/runc) - Reference implementation of OCI runtime
-
-## Runtime (Virtual Machine)
-
-* [hyperhq/runv](https://github.com/hyperhq/runv) - Hypervisor-based runtime for OCI
-
-## Bundle authoring
-
-* [kunalkushwaha/octool](https://github.com/kunalkushwaha/octool) - A config linter and validator.
-* [mrunalp/ocitools](https://github.com/mrunalp/ocitools) - A config generator.
-
-## Testing
-
-* [huawei-openlab/oct](https://github.com/huawei-openlab/oct) - Open Container Testing framework for OCI configuration and runtime
